backend/requirements.txt
django
djangorestframework
Pillow
requests